Move comon code from isodoc/*_convert.rb to isodoc/base_convert.rb
