+++ lib/isodoc/acme/pdf_convert.rb
@@ -1,6 +1,5 @@
+require_relative "base_convert"
 require "isodoc"
-require_relative "metadata"
-require "fileutils"
 
 module IsoDoc
   module Acme
@@ -9,12 +8,6 @@ module IsoDoc
     class PdfConvert < IsoDoc::PdfConvert
       def initialize(options)
         @libdir = File.dirname(__FILE__)
-        super
-      end
-
-      def convert1(docxml, filename, dir)
-        FileUtils.cp html_doc_path('logo.jpg'), File.join(@localdir, "logo.jpg")
-        @files_to_delete << File.join(@localdir, "logo.jpg")
         super
       end
 
@@ -35,10 +28,6 @@ module IsoDoc
         }
       end
 
-      def metadata_init(lang, script, labels)
-        @meta = Metadata.new(lang, script, labels)
-      end
-
       def googlefonts
         <<~HEAD.freeze
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i|Space+Mono:400,700" rel="stylesheet">
@@ -46,64 +35,11 @@ module IsoDoc
         HEAD
       end
 
-      def make_body(xml, docxml)
-        body_attr = { lang: "EN-US", link: "blue", vlink: "#954F72", "xml:lang": "EN-US", class: "container" }
-        xml.body **body_attr do |body|
-          make_body1(body, docxml)
-          make_body2(body, docxml)
-          make_body3(body, docxml)
-        end
-      end
-
       def html_toc(docxml)
         docxml
       end
 
-      def annex_name(annex, name, div)
-        div.h1 **{ class: "Annex" } do |t|
-          t << "#{get_anchors[annex['id']][:label]} "
-          t.br
-          t.b do |b|
-            name&.children&.each { |c2| parse(c2, b) }
-          end
-        end
-      end
-
-      def term_defs_boilerplate(div, source, term, preface)
-        if source.empty? && term.nil?
-          div << @no_terms_boilerplate
-        else
-          div << term_defs_boilerplate_cont(source, term)
-        end
-      end
-
-      def i18n_init(lang, script)
-        super
-        @annex_lbl = "Appendix"
-      end
-
-      def fileloc(loc)
-        File.join(File.dirname(__FILE__), loc)
-      end
-
-      def cleanup(docxml)
-        super
-        term_cleanup(docxml)
-      end
-
-      def term_cleanup(docxml)
-        docxml.xpath("//p[@class = 'Terms']").each do |d|
-          h2 = d.at("./preceding-sibling::*[@class = 'TermNum'][1]")
-          h2.add_child("&nbsp;")
-          h2.add_child(d.remove)
-        end
-        docxml
-      end
-
-      def info(isoxml, out)
-        @meta.security isoxml, out
-        super
-      end
+      include BaseConvert
     end
   end
 end
